create abstraction for querying the parsing context of analysis method
