+++ substratevm/src/com.oracle.svm.graal.hotspot.libgraal/src/com/oracle/svm/graal/hotspot/libgraal/LibGraalFeature.java
@@ -54,7 +54,7 @@ import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 import com.oracle.graal.pointsto.BigBang;
-import com.oracle.graal.pointsto.ReachabilityAnalysis;
+import com.oracle.graal.pointsto.meta.InvokeInfo;
 import org.graalvm.collections.EconomicMap;
 import org.graalvm.collections.MapCursor;
 import org.graalvm.compiler.code.DisassemblerProvider;
@@ -560,7 +560,7 @@ public final class LibGraalFeature implements com.oracle.svm.core.graal.GraalFea
             if (!isAllowedType(className)) {
                 disallowedTypes.add(className);
             }
-            for (ReachabilityAnalysis.InvokeInfo invoke : bb.getInvokes(m)) {
+            for (InvokeInfo invoke : bb.getInvokes(m)) {
                 for (AnalysisMethod callee : invoke.getPossibleCallees()) {
                     if (seen.add(callee)) {
                         todo.add(callee);
